// src/app/models/student.model.ts
export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  age: number;
  gender: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  country: string;
  department: string;
  level: number;
  GPA: number;
  enrollmentDate: string;
  isActive: boolean;
}